requirements.txt
sentence-transformers
psycopg2-binary
pgvector
fastapi